requirements.txt
langchain-groq>=0.0.5
python-dotenv>=0.19.0
Pillow>=10.0.0
requests>=2.31.0
python-jose>=3.3.0

langchain_community
langchain

transformers
torch
accelerate

gTTS

flask
werkzeug
gunicorn
